refactor split app
