+++ db.txt
@@ -1,3 +1,16 @@
+src/
+  App.jsx
+  lib/
+    supabaseClient.js
+  components/
+    Auth.jsx
+  features/
+    money/
+      MoneyApp.jsx
+
+
+
+
 -- Table
 create table if not exists public.expenses (
   id uuid primary key default gen_random_uuid(),
